patient&doctor correction

+++ src/app/modules/Patient/patient.service.ts
@@ -96,27 +96,48 @@ const deletePatientService = async (id: string): Promise<Patient> => {
     throw new ApiError(404, "id does not exist");
   }
 
-  const result = await prisma.$transaction(async (tx) => {
 
+
+  const appointmentExist = await prisma.appointment.findMany({
+    where: {
+      patientId:id
+    },
+  });
+
+  //check if appointmentExist
+  if(appointmentExist.length > 0){
+    throw new ApiError(409, 'This patientId is associated with Appointment');
+  }
+
+  
+  const patientHealthDataExist = await prisma.patientHealthData.findUnique({
+    where: {
+      patientId: patient.id
+    }
+  })
+
+
+  const result = await prisma.$transaction(async (tx) => {
     //query-01 delete medicalReport
     await tx.medicalReport.deleteMany({
       where: {
-        patientId: patient.id
-      }
-    })
+        patientId: patient.id,
+      },
+    });
 
-     //query-02 delete patientHealthData
-     await tx.patientHealthData.delete({
-      where: {
-        patientId: patient.id
-      }
-    })
-
+    //query-02 delete patientHealthData
+    if (patientHealthDataExist) {
+      await tx.patientHealthData.delete({
+        where: {
+          patientId: patient.id,
+        },
+      });
+    }
 
     //query-03 patient delete
-    const patientDeletedData =await tx.patient.delete({
+    const patientDeletedData = await tx.patient.delete({
       where: {
-        id
+        id,
       },
     });
 
@@ -130,7 +151,8 @@ const deletePatientService = async (id: string): Promise<Patient> => {
     return patientDeletedData;
   });
 
-  return result;
+  return result
+
 };
